fix: improve new device logging, resolves #90
